+++ src/extent.js
@@ -1,3 +1,5 @@
+import scaleCoordinate from './scaleCoordinate.js';
+
 export default function (elements) {
 	const extent = {
 		xmin: 0,
@@ -9,7 +11,9 @@ export default function (elements) {
 	};
 
 	let isFirstCoordinate = true;
+	let needs2ndPass = false;
 
+	// 1st pass.
 	elements.forEach((element) => {
 		if (
 			element.type === 'arrow' ||
@@ -50,6 +54,7 @@ export default function (elements) {
 				}
 			}));
 		} else {
+			// The comments bellow are also valid for the 2nd pass.
 			// The extent isn't calculated correctly for cylinders, their extent should be transformationVector * radius.
 			// The calculated extent for polyhedrons is approximated.
 			// The calculated extent for tubes is approximated.
@@ -89,49 +94,54 @@ export default function (elements) {
 							extent.zmax = coordinate[0][2] + radius;
 						}
 					}
+				} else {
+					needs2ndPass = true;
 				}
 			}));
 		}
 	});
 
-	// The comments bellow are also valid for ymin/ymax and zmin/zmax.
-	if (extent.xmin > extent.xmax) {
-		// xmin should be smaller than xmax, swap them.
-		[extent.xmin, extent.xmax] = [extent.xmax, extent.xmin];
-	} else if (extent.xmin === extent.xmax) {
-		// There is only one non-scaled coordinate, so we transform this extent
-		// into another, otherwise nothing would appear in the screen.
-		if (extent.xmin < 0) {
-			[extent.xmin, extent.xmax] = [2 * extent.xmin, 0];
-		} else if (extent.xmin > 0) {
-			[extent.xmin, extent.xmax] = [0, 2 * extent.xmin];
-		} else { // xmin === xmax === 0
-			[extent.xmin, extent.xmax] = [-1, 1];
-		}
-	}
+	if (needs2ndPass) {
+		// 2nd pass, necesary for primitives with scaled coordinates and
+		// radius > extent.
+		elements.forEach((element) => {
+			if (
+				// arrows, cuboids, lines and polygons have no radius.
+				element.type !== 'arrow' &&
+				element.type !== 'cuboid' &&
+				element.type !== 'line' &&
+				element.type !== 'polygon'
+			) {
+				const radius = element.radius ?? element.pointSize ?? element.edgeLength ?? 1;
 
-	if (extent.ymin > extent.ymax) {
-		[extent.ymin, extent.ymax] = [extent.ymax, extent.ymin];
-	} else if (extent.ymin === extent.ymax) {
-		if (extent.ymin < 0) {
-			[extent.ymin, extent.ymax] = [2 * extent.ymin, 0];
-		} else if (extent.ymin > 0) {
-			[extent.ymin, extent.ymax] = [0, 2 * extent.ymin];
-		} else {
-			[extent.ymin, extent.ymax] = [-1, 1];
-		}
-	}
+				element.coords.forEach(((coordinate) => {
+					if (!coordinate[0]) {
+						coordinate[0] = scaleCoordinate(coordinate[1], extent);
 
-	if (extent.zmin > extent.zmax) {
-		[extent.zmin, extent.zmax] = [extent.zmax, extent.zmin];
-	} else if (extent.zmin === extent.zmax) {
-		if (extent.zmin < 0) {
-			[extent.zmin, extent.zmax] = [2 * extent.zmin, 0];
-		} else if (extent.zmin > 0) {
-			[extent.zmin, extent.zmax] = [0, 2 * extent.zmin];
-		} else {
-			[extent.zmin, extent.zmax] = [-1, 1];
-		}
+						if (coordinate[0][0] - radius < extent.xmin) {
+							extent.xmin = coordinate[0][0] - radius;
+						}
+						if (coordinate[0][0] + radius > extent.xmax) {
+							extent.xmax = coordinate[0][0] + radius;
+						}
+
+						if (coordinate[0][1] - radius < extent.ymin) {
+							extent.ymin = coordinate[0][1] - radius;
+						}
+						if (coordinate[0][1] + radius > extent.ymax) {
+							extent.ymax = coordinate[0][1] + radius;
+						}
+
+						if (coordinate[0][2] - radius < extent.zmin) {
+							extent.zmin = coordinate[0][2] - radius;
+						}
+						if (coordinate[0][2] + radius > extent.zmax) {
+							extent.zmax = coordinate[0][2] + radius;
+						}
+					}
+				}));
+			}
+		});
 	}
 
 	return extent;