+++ CHANGES.md
@@ -10,6 +10,9 @@ Documentation:
 Internals:
 - Update minify to version 8
 
+Bug fixes:
+- Correctly calculate the bounding box for primitives with scaled coordinates and radius bigger than the bounding box (#60)
+
 1.1.0
 -----